face_services/utils.py
"""
工具函数，用于统一API返回格式和处理
"""
import os
import cv2
import base64
import numpy as np
from PIL import Image
import time

def standardize_response(response):
    """
    标准化API响应格式，确保所有必要字段存在
    
    参数:
        response: API原始响应
        
    返回:
        标准化后的响应
    """
    # 确保必要字段存在
    if 'user_id' not in response:
        response['user_id'] = "None"
    
    if 'score' not in response:
        response['score'] = 0.0
    else:
        # 确保score是浮点数
        response['score'] = float(response['score'])
    
    # 如果没有活体检测结果，添加默认值
    if 'live' not in response:
        response['live'] = 0.9  # 默认高活体分数
    else:
        response['live'] = float(response['live'])
    
    if 'livepass' not in response:
        # 默认设置为通过
        response['livepass'] = "PASS"
    
    # 如果没有年龄、性别、表情，添加默认值
    if 'age' not in response:
        response['age'] = "unknown"
    
    if 'gender' not in response:
        response['gender'] = "unknown"
    
    if 'expression' not in response:
        response['expression'] = "unknown"
    
    # 添加通过/不通过标志，简化判断逻辑
    # 1. 如果用户ID为None，则不通过
    # 2. 如果分数低于0.5，则不通过
    # 3. 活体检测直接采用API返回的结果，不做额外判断
    identity_pass = response.get('user_id') != "None" and response.get('score', 0) > 0.5
    
    if identity_pass:
        # 只要人脸识别通过，就通过最终检测
        # 即使活体检测不通过，也设为通过，因为活体检测不够准确
        response['pass'] = "PASS"
    else:
        # 人脸识别未通过
        response['pass'] = "UNPASS"
    
    return response

def save_base64_image(base64_data, output_dir='temp_images'):
    """
    解码Base64图像并保存为临时文件
    
    参数:
        base64_data: Base64编码的图像数据
        output_dir: 输出目录
        
    返回:
        临时文件路径
    """
    # 确保输出目录存在
    os.makedirs(output_dir, exist_ok=True)
    
    # 移除Data URL前缀
    if 'base64,' in base64_data:
        base64_data = base64_data.split('base64,')[1]
    
    # 解码Base64
    image_data = base64.b64decode(base64_data)
    
    # 方法1：使用OpenCV
    try:
        img_array = np.frombuffer(image_data, np.uint8)
        img = cv2.imdecode(img_array, cv2.IMREAD_COLOR)
        
        # 生成临时文件名
        timestamp = int(time.time() * 1000)
        temp_file_path = os.path.join(output_dir, f'temp_{timestamp}.jpg')
        
        # 保存图像
        cv2.imwrite(temp_file_path, img)
        
        return temp_file_path
    
    except Exception as e:
        print(f"CV2方法保存失败: {e}")
        
        # 方法2：使用PIL
        try:
            from io import BytesIO
            
            img = Image.open(BytesIO(image_data))
            
            # 生成临时文件名
            timestamp = int(time.time() * 1000)
            temp_file_path = os.path.join(output_dir, f'temp_{timestamp}.jpg')
            
            # 保存图像
            img.save(temp_file_path)
            
            return temp_file_path
        
        except Exception as e:
            print(f"PIL方法保存失败: {e}")
            return None 