requirements.txt
# 人脸识别核心依赖
face-recognition==1.3.0
facenet-pytorch==2.5.2
torch==1.12.1
torchvision==0.13.1
numpy==1.22.4
Pillow>=9.0.0

# API服务
baidu-aip==4.16.9

# Web应用
Flask==2.2.3
pandas==1.5.3
openpyxl==3.1.2

# 其他工具
matplotlib==3.7.1
opencv-python==4.7.0.72
scipy==1.10.1